added label to canvas
