+++ lib/d/json_textform/models/Schema.dart
@@ -9,7 +9,7 @@ enum WidgetType {
   unknown,
   select,
   checkbox,
-  signature
+  canvas
 }
 
 class Schema {
